Раскидал по папкам и немного фиксов
