+++ VKApp/Friends/Controls/FriendsSearchBar.swift
