+++ VKApp/Friends/Controls/LikeCounterControl.swift
