+++ VKApp/Friends/Controls/SorterBarControl.swift
