+++ VKApp/Friends/PhotoViewer/FullPhotoViewController.swift
