+++ VKApp/Friends/PhotoViewer/PhotoCollectionViewCell.swift
