+++ VKApp/Friends/VKApp/Base.lproj/Friends.storyboard
@@ -4,7 +4,6 @@
     <dependencies>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="collection view cell content view" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
@@ -81,7 +80,7 @@
                                                 </imageView>
                                                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="f1b-w3-9r9">
                                                     <rect key="frame" x="90" y="19.5" width="273" height="31.5"/>
-                                                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                                     <accessibility key="accessibilityConfiguration" identifier="user">
                                                         <accessibilityTraits key="traits" staticText="YES"/>
                                                     </accessibility>
@@ -125,67 +124,6 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="BUJ-5e-IrT" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="3173.913043478261" y="-198.21428571428569"/>
-        </scene>
-        <!--Collection View Controller-->
-        <scene sceneID="JjD-nT-Ccb">
-            <objects>
-                <collectionViewController id="mjJ-RJ-yWT" sceneMemberID="viewController">
-                    <collectionView key="view" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="prototypes" id="HTL-je-0Wb">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                        <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="uQx-p6-VM7">
-                            <size key="itemSize" width="160" height="160"/>
-                            <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                            <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                            <inset key="sectionInset" minX="20" minY="20" maxX="20" maxY="20"/>
-                        </collectionViewFlowLayout>
-                        <cells>
-                            <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="photoCell" id="FiS-0p-7kr" customClass="PhotoCollectionViewCell" customModule="VKApp" customModuleProvider="target">
-                                <rect key="frame" x="20" y="20" width="160" height="160"/>
-                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                                <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="Ow7-JZ-mKJ">
-                                    <rect key="frame" x="0.0" y="0.0" width="160" height="160"/>
-                                    <autoresizingMask key="autoresizingMask"/>
-                                    <subviews>
-                                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="gvU-te-5K1">
-                                            <rect key="frame" x="15" y="15" width="130" height="130"/>
-                                        </imageView>
-                                        <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="2m8-dT-XK1" customClass="LikeCounterControl" customModule="VKApp" customModuleProvider="target">
-                                            <rect key="frame" x="55" y="135" width="50" height="25"/>
-                                            <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                                            <constraints>
-                                                <constraint firstAttribute="height" constant="25" id="Wn1-dW-V43"/>
-                                                <constraint firstAttribute="width" constant="50" id="k8R-L8-rrQ"/>
-                                            </constraints>
-                                        </view>
-                                    </subviews>
-                                    <constraints>
-                                        <constraint firstAttribute="trailing" secondItem="gvU-te-5K1" secondAttribute="trailing" constant="15" id="5qc-sV-bnr"/>
-                                        <constraint firstItem="2m8-dT-XK1" firstAttribute="centerX" secondItem="Ow7-JZ-mKJ" secondAttribute="centerX" id="Aaj-n6-ZWc"/>
-                                        <constraint firstItem="gvU-te-5K1" firstAttribute="top" secondItem="Ow7-JZ-mKJ" secondAttribute="top" constant="15" id="U7p-CH-eoU"/>
-                                        <constraint firstAttribute="bottom" secondItem="2m8-dT-XK1" secondAttribute="bottom" id="c9d-xZ-BR1"/>
-                                        <constraint firstAttribute="bottom" secondItem="gvU-te-5K1" secondAttribute="bottom" constant="15" id="pfj-pM-t6c"/>
-                                        <constraint firstItem="gvU-te-5K1" firstAttribute="leading" secondItem="Ow7-JZ-mKJ" secondAttribute="leading" constant="15" id="v2Z-NO-ZXM"/>
-                                    </constraints>
-                                </collectionViewCellContentView>
-                                <size key="customSize" width="160" height="160"/>
-                                <connections>
-                                    <outlet property="likeControl" destination="2m8-dT-XK1" id="87T-ni-sRa"/>
-                                    <outlet property="photoImageView" destination="gvU-te-5K1" id="JL5-Pl-SH5"/>
-                                </connections>
-                            </collectionViewCell>
-                        </cells>
-                        <connections>
-                            <outlet property="dataSource" destination="mjJ-RJ-yWT" id="Vh0-W4-8Ly"/>
-                            <outlet property="delegate" destination="mjJ-RJ-yWT" id="LMg-be-1z1"/>
-                        </connections>
-                    </collectionView>
-                    <navigationItem key="navigationItem" id="Agr-aY-dJD"/>
-                </collectionViewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="ilL-v9-Rjl" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="4090" y="-892"/>
         </scene>
         <!--Photo List View Controller-->
         <scene sceneID="YdO-e0-0pq">
